feat: Show success page to admin and pass selected membership duration from MembershipDurationPage to CameraPage
Replaced the dialog with a page to show success to the admin after confirming payment. The SuccessPage is navigated to upon successful payment confirmation, displaying relevant information about the newly added member and the duration of membership.

Additionally, modified MembershipDurationPage to pass the selected number of months to the CameraPage widget by adding the 'months' parameter to the CameraPage constructor. This allows the CameraPage to access and utilize the selected membership duration for further processing.